// reolio/src/generate/schema/http/models/MediaGet200Response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * My Project
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from "../runtime";
import type { MediaGet200ResponseMediaInner } from "./MediaGet200ResponseMediaInner";
import {
  MediaGet200ResponseMediaInnerFromJSON,
  MediaGet200ResponseMediaInnerFromJSONTyped,
  MediaGet200ResponseMediaInnerToJSON,
} from "./MediaGet200ResponseMediaInner";

/**
 *
 * @export
 * @interface MediaGet200Response
 */
export interface MediaGet200Response {
  /**
   *
   * @type {number}
   * @memberof MediaGet200Response
   */
  totalCount: number;
  /**
   *
   * @type {Array<MediaGet200ResponseMediaInner>}
   * @memberof MediaGet200Response
   */
  media: Array<MediaGet200ResponseMediaInner>;
}

/**
 * Check if a given object implements the MediaGet200Response interface.
 */
export function instanceOfMediaGet200Response(value: object): boolean {
  if (!("totalCount" in value)) return false;
  if (!("media" in value)) return false;
  return true;
}

export function MediaGet200ResponseFromJSON(json: any): MediaGet200Response {
  return MediaGet200ResponseFromJSONTyped(json, false);
}

export function MediaGet200ResponseFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean,
): MediaGet200Response {
  if (json == null) {
    return json;
  }
  return {
    totalCount: json["total_count"],
    media: (json["media"] as Array<any>).map(
      MediaGet200ResponseMediaInnerFromJSON,
    ),
  };
}

export function MediaGet200ResponseToJSON(
  value?: MediaGet200Response | null,
): any {
  if (value == null) {
    return value;
  }
  return {
    total_count: value["totalCount"],
    media: (value["media"] as Array<any>).map(
      MediaGet200ResponseMediaInnerToJSON,
    ),
  };
}