// reolio/src/generate/schema/http/models/MediaMediumIdGet200Response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * My Project
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from "../runtime";
/**
 *
 * @export
 * @interface MediaMediumIdGet200Response
 */
export interface MediaMediumIdGet200Response {
  /**
   *
   * @type {string}
   * @memberof MediaMediumIdGet200Response
   */
  mediumId: string;
  /**
   *
   * @type {string}
   * @memberof MediaMediumIdGet200Response
   */
  thumbnailUrl: string;
  /**
   *
   * @type {string}
   * @memberof MediaMediumIdGet200Response
   */
  downloadUrl: string;
  /**
   *
   * @type {string}
   * @memberof MediaMediumIdGet200Response
   */
  fileType: string;
}

/**
 * Check if a given object implements the MediaMediumIdGet200Response interface.
 */
export function instanceOfMediaMediumIdGet200Response(value: object): boolean {
  if (!("mediumId" in value)) return false;
  if (!("thumbnailUrl" in value)) return false;
  if (!("downloadUrl" in value)) return false;
  if (!("fileType" in value)) return false;
  return true;
}

export function MediaMediumIdGet200ResponseFromJSON(
  json: any,
): MediaMediumIdGet200Response {
  return MediaMediumIdGet200ResponseFromJSONTyped(json, false);
}

export function MediaMediumIdGet200ResponseFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean,
): MediaMediumIdGet200Response {
  if (json == null) {
    return json;
  }
  return {
    mediumId: json["medium_id"],
    thumbnailUrl: json["thumbnail_url"],
    downloadUrl: json["download_url"],
    fileType: json["file_type"],
  };
}

export function MediaMediumIdGet200ResponseToJSON(
  value?: MediaMediumIdGet200Response | null,
): any {
  if (value == null) {
    return value;
  }
  return {
    medium_id: value["mediumId"],
    thumbnail_url: value["thumbnailUrl"],
    download_url: value["downloadUrl"],
    file_type: value["fileType"],
  };
}