// reolio/src/generate/schema/http/models/MediaPost200Response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * My Project
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from "../runtime";
/**
 *
 * @export
 * @interface MediaPost200Response
 */
export interface MediaPost200Response {
  /**
   *
   * @type {string}
   * @memberof MediaPost200Response
   */
  uploadPresignedUrl: string;
}

/**
 * Check if a given object implements the MediaPost200Response interface.
 */
export function instanceOfMediaPost200Response(value: object): boolean {
  if (!("uploadPresignedUrl" in value)) return false;
  return true;
}

export function MediaPost200ResponseFromJSON(json: any): MediaPost200Response {
  return MediaPost200ResponseFromJSONTyped(json, false);
}

export function MediaPost200ResponseFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean,
): MediaPost200Response {
  if (json == null) {
    return json;
  }
  return {
    uploadPresignedUrl: json["upload_presigned_url"],
  };
}

export function MediaPost200ResponseToJSON(
  value?: MediaPost200Response | null,
): any {
  if (value == null) {
    return value;
  }
  return {
    upload_presigned_url: value["uploadPresignedUrl"],
  };
}